@@ -50,60 +54,58 @@
     "benchmark": "2.1.4",
     "bluebird": "3.5.1",
     "cli-table2": "0.2.0",
-    "eslint": "^4.13.0",
+    "eslint": "^4.18.2",
     "eslint-config-rapid7": "^2.11.1",
     "eslint-friendly-formatter": "3.0.0",
-    "eslint-loader": "1.9.0",
-    "fast-equals": "1.0.5",
-    "fast-memoize": "2.2.8",
-    "flow-babel-webpack-plugin": "1.1.0",
-    "flow-bin": "^0.61.0",
-    "html-webpack-plugin": "2.30.1",
+    "eslint-loader": "2.0.0",
+    "fast-equals": "1.2.1",
+    "fast-memoize": "2.3.0",
+    "flow-babel-webpack-plugin": "1.1.1",
+    "flow-bin": "^0.67.1",
+    "html-webpack-plugin": "3.0.6",
     "husky": "0.14.3",
     "in-publish": "2.0.0",
     "ink-docstrap": "1.3.2",
     "jsdoc": "3.5.5",
-    "jsdoc-babel": "0.3.0",
-    "jsdom": "^11.5.1",
+    "jsdoc-babel": "0.4.0",
+    "jsdom": "^11.6.2",
     "jsdom-global": "3.0.2",
-    "lodash": "4.17.4",
+    "lodash": "4.17.5",
     "lru-memoize": "1.0.2",
-    "memoizee": "^0.4.11",
+    "mem": "^3.0.0",
+    "memoizee": "^0.4.12",
     "memoizerific": "1.11.2",
-    "nyc": "11.3.0",
+    "nyc": "11.4.1",
     "optimize-js-plugin": "0.0.4",
-    "ora": "1.3.0",
-    "prop-types": "15.6.0",
+    "ora": "2.0.0",
+    "prop-types": "15.6.1",
     "q": "1.5.1",
     "ramda": "0.25.0",
     "react": "16.2.0",
     "react-dom": "16.2.0",
     "rimraf": "2.6.2",
-    "sinon": "^4.1.3",
+    "rollup": "0.56.5",
+    "rollup-plugin-babel": "3.0.3",
+    "rollup-plugin-commonjs": "9.0.0",
+    "rollup-plugin-node-resolve": "3.2.0",
+    "rollup-plugin-uglify": "3.0.0",
+    "sinon": "^4.4.2",
     "underscore": "1.8.3",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.9.7"
+    "webpack": "^4.1.1",
+    "webpack-cli": "2.0.11",
+    "webpack-dev-server": "^3.1.1"
   },
   "homepage": "https://github.com/planttheidea/moize#readme",
   "keywords": [
-    "memoize",
-    "memoized",
-    "memoizer",
-    "memoization",
-    "memo",
-    "memcache",
-    "memcached",
     "cache",
-    "cached",
-    "async",
-    "storage",
-    "memory",
+    "expire",
+    "lru",
+    "memoize",
+    "memoization",
     "optimize",
     "performance",
-    "lru",
-    "ttl",
-    "expire",
-    "promise"
+    "promise",
+    "ttl"
   ],
   "license": "MIT",
   "main": "lib/index.js",
