@@ -1,26 +1,26 @@
 declare module 'moize' {
-  declare type PromiseLike<R> = {
-    then<U>(onFulfill?: (value: R) => Promise<U> | U, onReject?: (error: any) => Promise<U> | U): Promise<U>
+  declare type Cache = {
+    keys: Array<Array<any>>,
+    size: number,
+    values: Array<any>
   };
 
-  declare type PromiseLibrary<T> = {
-    (callback: (resolve: (r?: T | PromiseLike<T>) => void, reject: (e?: any) => void) => void): PromiseLike<T>,
-    reject: (err: Error) => any,
-    resolve: (v: T) => any
-  };
-
-  declare type Config = {
+  declare type Options = {
     equals?: (firstValue: any, secondValue: any) => boolean,
+    isDeepEqual?: boolean,
     isPromise?: boolean,
     isReact?: boolean,
+    isSerialized?: boolean,
     maxAge?: number,
     maxArgs?: number,
     maxSize?: number,
+    onCacheAdd?: (cache: Cache) => void,
+    onCacheChange?: (cache: Cache) => void,
+    onCacheHit?: (cache: Cache) => void,
     onExpire?: (key: any) => void,
-    promiseLibrary?: PromiseLibrary<any>,
-    serialize?: boolean,
-    serializeFunctions?: boolean,
+    profileName?: string,
     serializer?: (...args: any[]) => any,
+    shouldSerializeFunctions?: boolean,
     transformArgs?: (args: any[]) => any[],
     updateExpire?: boolean
   };
@@ -29,18 +29,22 @@ declare module 'moize' {
 
   declare type Moizer<T> = (t: T) => T;
 
-  declare module.exports: {
-    (config: Config): Fn,
-    (fn: Fn, config?: Config): Fn,
+  declare export function collectStats(): void;
 
-    maxAge<T>(a: number): (t: T, c?: Config) => T,
-    maxArgs<T>(a: number): (t: T, c?: Config) => T,
-    maxSize<T>(a: number): (t: T, c?: Config) => T,
-    promise<T>(t: T, c?: Config): T,
-    react<T>(t: T, c?: Config): T,
-    reactSimple<T>(t: T, c?: Config): T,
-    serialize<T>(t: T, c?: Config): T,
-    simple<T>(t: T, c?: Config): T,
-    compose<T>(...fns: Array<Moizer<T>>): Moizer<T>
+  declare export default {
+    (options: Options): Fn,
+    (fn: Fn, options?: Options): Fn,
+
+    compose<T>(...fns: Array<Moizer<T>>): Moizer<T>,
+    deep<T>(t: T, c?: Options): T,
+    isMoized<T>(t: T): boolean,
+    maxAge<T>(a: number): (t: T, c?: Options) => T,
+    maxArgs<T>(a: number): (t: T, c?: Options) => T,
+    maxSize<T>(a: number): (t: T, c?: Options) => T,
+    promise<T>(t: T, c?: Options): T,
+    react<T>(t: T, c?: Options): T,
+    reactSimple<T>(t: T, c?: Options): T,
+    serialize<T>(t: T, c?: Options): T,
+    simple<T>(t: T, c?: Options): T
   };
 }
