@@ -1,13 +1,29 @@
 import 'babel-polyfill';
 
-import isEqual from 'lodash/isEqual';
 import Bluebird from 'bluebird';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import {render} from 'react-dom';
 import memoizee from 'memoizee';
 
-import moize from '../src';
+import moize, {collectStats} from '../src';
+
+const div = document.createElement('div');
+
+div.id = 'app-container';
+div.style.backgroundColor = '#1d1d1d';
+div.style.boxSizing = 'border-box';
+div.style.color = '#d5d5d5';
+div.style.height = '100vh';
+div.style.padding = '15px';
+div.style.width = '100vw';
+
+document.body.style.margin = 0;
+document.body.style.padding = 0;
+
+document.body.appendChild(div);
+
+collectStats();
 
 console.group('standard');
 
@@ -32,6 +48,8 @@ console.log(memoized.cache);
 console.log('has true', memoized.has([foo, bar]));
 console.log('has false', memoized.has([foo, 'baz']));
 
+console.log(memoized.getStats());
+
 console.groupEnd('standard');
 
 console.group('maxArgs');
@@ -43,17 +61,15 @@ memoizedMax(foo, 'baz');
 
 console.groupEnd('maxArgs');
 
-console.group('custom - deep equals');
+console.group('deep equals');
 
 const deepEqualMethod = ({one, two}) => {
-  console.log('custom equal method fired', one, two);
+  console.log('deep equalfired', one, two);
 
   return [one, two];
 };
 
-const deepEqualMemoized = moize(deepEqualMethod, {
-  equals: isEqual
-});
+const deepEqualMemoized = moize.deep(deepEqualMethod);
 
 deepEqualMemoized({one: 1, two: 2});
 deepEqualMemoized({one: 2, two: 1});
@@ -64,14 +80,35 @@ console.log(deepEqualMemoized.cache);
 console.log('has deep true', deepEqualMemoized.has([{one: 1, two: 2}]));
 console.log('has deep false', deepEqualMemoized.has([{one: 1, two: 3}]));
 
-console.groupEnd('custom - deep equals');
+console.groupEnd('deep equals');
+
+console.group('serialize');
+
+const serializeMethod = ({one, two}) => {
+  console.log('serialize fired', one, two);
+
+  return [one, two];
+};
+
+const serializeMemized = moize.serialize(serializeMethod);
+
+serializeMemized({one: 1, two: 2});
+serializeMemized({one: 2, two: 1});
+serializeMemized({one: 1, two: 2});
+serializeMemized({one: 1, two: 2});
+
+console.log(serializeMemized.cache);
+console.log('has serialized true', serializeMemized.has([{one: 1, two: 2}]));
+console.log('has serialized false', serializeMemized.has([{one: 1, two: 3}]));
+
+console.groupEnd('serialize');
 
 console.group('promise');
 
 const promiseMethod = (number, otherNumber) => {
   console.log('promise method fired', number);
 
-  return new Promise((resolve) => {
+  return new Bluebird((resolve) => {
     resolve(number * otherNumber);
   });
 };
@@ -89,9 +126,10 @@ const promiseMethodRejected = (number) => {
 const memoizedPromise = moize(promiseMethod, {
   isPromise: true
 });
-const memoizedPromiseRejected = moize({isPromise: true})({promiseLibrary: Bluebird})(promiseMethodRejected);
+const memoizedPromiseRejected = moize({isPromise: true, profileName: 'rejected promise'})(promiseMethodRejected);
 
 console.log('curried options', memoizedPromiseRejected.options);
+console.log('curried options under the hood', memoizedPromiseRejected._microMemoizeOptions);
 
 memoizedPromiseRejected(3)
   .then((foo) => {
@@ -99,6 +137,9 @@ memoizedPromiseRejected(3)
   })
   .catch((bar) => {
     console.error(bar);
+  })
+  .finally(() => {
+    console.log(memoizedPromiseRejected.keys());
   });
 
 memoizedPromiseRejected(3)
@@ -107,6 +148,9 @@ memoizedPromiseRejected(3)
   })
   .catch((bar) => {
     console.error(bar);
+  })
+  .finally(() => {
+    console.log(memoizedPromiseRejected.keys());
   });
 
 memoizedPromiseRejected(3)
@@ -115,6 +159,9 @@ memoizedPromiseRejected(3)
   })
   .catch((bar) => {
     console.error(bar);
+  })
+  .finally(() => {
+    console.log(memoizedPromiseRejected.keys());
   });
 
 // get result
@@ -128,6 +175,28 @@ memoizedPromise(2, 2).then((value) => {
 });
 
 console.log(memoizedPromise.keys());
+
+const otherPromiseMethod = (number) => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve(number * 2);
+    }, 1000);
+  });
+};
+
+const memoizedOtherPromise = moize.promise(otherPromiseMethod, {
+  maxAge: 1500,
+  onCacheHit(cache) {
+    console.log('must have resolved!', cache);
+  },
+  onExpire() {
+    console.log('updated promise expired');
+  }
+});
+
+memoizedOtherPromise(4).then((number) => {
+  console.log('i should be 8', number);
+});
 
 console.groupEnd('promise');
 
@@ -203,11 +272,13 @@ Foo.defaultProps = {
   bar: 'default'
 };
 
-const MemoizedFoo = moize.react(Foo);
-const SimpleMemoizedFoo = moize.reactSimple(Foo);
+const MemoizedFoo = moize.react(Foo, {isDeepEqual: true});
+const SimpleMemoizedFoo = moize.reactSimple(Foo, {profileName: 'SimpleMemoizedFoo'});
+const LimitedMemoizedFoo = moize.compose()(Foo);
 
-console.log('MemoizedFoo', MemoizedFoo.options);
-console.log('SimpleMemoizedFoo', SimpleMemoizedFoo.options);
+console.log('MemoizedFoo', MemoizedFoo.options, MemoizedFoo._microMemoizeOptions);
+console.log('SimpleMemoizedFoo', SimpleMemoizedFoo.options, SimpleMemoizedFoo._microMemoizeOptions);
+console.log('LimitedMemoizedFoo', LimitedMemoizedFoo.options, LimitedMemoizedFoo._microMemoizeOptions);
 
 console.log('MemoizedFoo cache', MemoizedFoo.cache);
 
@@ -226,15 +297,17 @@ const expiringMemoized = moize(method, {
       if (count !== 0) {
         console.log(
           'Expired! This is the last time I will fire, and this should be empty:',
-          expiringMemoized.cache.expirations
+          expiringMemoized.expirationsSnapshot
         );
+
+        console.log(moize.getStats());
 
         return true;
       }
 
       console.log(
         'Expired! I will now reset the expiration, but this should be empty:',
-        expiringMemoized.cache.expirations
+        expiringMemoized.expirationsSnapshot
       );
 
       count++;
@@ -253,7 +326,11 @@ expiringMemoized(foo, bar);
 expiringMemoized(foo, bar);
 expiringMemoized(foo, bar);
 
+console.log('existing expirations', expiringMemoized.expirationsSnapshot);
+
 console.groupEnd('expiration');
+
+console.log(moize.getStats());
 
 const HEADER_STYLE = {
   margin: 0
@@ -289,19 +366,4 @@ class App extends Component {
   }
 }
 
-const div = document.createElement('div');
-
-div.id = 'app-container';
-div.style.backgroundColor = '#1d1d1d';
-div.style.boxSizing = 'border-box';
-div.style.color = '#d5d5d5';
-div.style.height = '100vh';
-div.style.padding = '15px';
-div.style.width = '100vw';
-
-document.body.style.margin = 0;
-document.body.style.padding = 0;
-
 render(<App />, div);
-
-document.body.appendChild(div);
