@@ -1,117 +1,128 @@
 // @flow
 
-// cache
-import Cache from './Cache';
+// external dependencies
+import memoize from 'micro-memoize';
 
 // constants
-import {INVALID_FIRST_PARAMETER_ERROR} from './constants';
+import {DEFAULT_OPTIONS} from './constants';
+
+// instance
+import {augmentMoizeInstance} from './instance';
+
+// max age
+import {getMaxAgeOptions} from './maxAge';
+
+// options
+import {createOnCacheOperation, getIsEqual, getTransformKey} from './options';
+
+// stats
+import {collectStats, getDefaultProfileName, getStats, getStatsOptions, statsCache} from './stats';
 
 // types
-import type {Options} from './types';
+import type {Expiration, MicroMemoizeOptions, Options} from './types';
 
 // utils
-import {
-  compose,
-  createAddPropertiesToFunction,
-  createCurriableOptionMethod,
-  createGetCacheKey,
-  createSetNewCachedValue,
-  getDefaultedOptions,
-  isFunction,
-  isPlainObject
-} from './utils';
+import {combine, compose, mergeOptions} from './utils';
 
 /**
  * @module moize
  */
 
+export {collectStats};
+
 /**
  * @function moize
  *
  * @description
- * store cached values returned from calling method with arguments to avoid reprocessing data from same arguments
+ * memoize a function based its arguments passed, potentially improving runtime performance
  *
  * @example
  * import moize from 'moize';
  *
  * // standard implementation
- * const fn = (foo, bar) => {
- *  return `${foo} ${bar}`;
- * };
+ * const fn = (foo, bar) => `${foo} ${bar}`;
  * const memoizedFn = moize(fn);
  *
  * // implementation with options
- * const fn = async (id) => {
- *  return get(`http://foo.com/${id}`);
- * };
- * const memoizedFn = moize(fn, {
- *  isPromise: true,
- *  maxSize: 5
- * });
+ * const fn = async (id) => get(`http://foo.com/${id}`);
+ * const memoizedFn = moize(fn, {isPromise: true, maxSize: 5});
  *
- * @param {function} functionOrComposableOptions method to memoize
- * @param {Options} [passedOptions={}] options to customize how the caching is handled
- * @param {boolean} [passedOptions.isPromise=false] is the function return expected to be a promise to resolve
- * @param {number} [passedOptions.maxAge=Infinity] the maximum age the value should persist in cache
- * @param {number} [passedOptions.maxArgs=Infinity] the maximum number of arguments to be used in serializing the keys
- * @param {number} [passedOptions.maxSize=Infinity] the maximum size of the cache to retain
- * @param {function} [passedOptions.promiseLibrary=Promise] promise library to use for resolution / rejection
- * @param {function} [passedOptions.serializeFunctions=false] should function parameters be serialized as well
- * @param {function} [passedOptions.serializer] method to serialize arguments with for cache storage
- * @returns {Function} higher-order function which either returns from cache or newly-computed value
+ * // implementation with convenience methods
+ * const Foo = ({foo}) => <div>{foo}</div>;
+ * const MemoizedFoo = moize.react(Foo);
+ *
+ * @param {function|Options} fn the function to memoized, or a list of options when currying
+ * @param {Options} [options=DEFAULT_OPTIONS] the options to apply
+ * @returns {function} the memoized function
  */
-const moize: Function = (functionOrComposableOptions: Function | Object, passedOptions: Object = {}): Function => {
-  if (isPlainObject(functionOrComposableOptions)) {
-    return function(fnOrOptions: Function | Object, otherOptions: Object = {}): Function {
-      if (isPlainObject(fnOrOptions)) {
-        return moize({
-          // $FlowIgnore functionOrComposableOptions is object of options
-          ...functionOrComposableOptions,
-          // $FlowIgnore fnOrOptions is object of options
-          ...fnOrOptions
-        });
-      }
+function moize(fn: Function | Options, options: Options = DEFAULT_OPTIONS): Function {
+  if (fn.isMoized) {
+    // $FlowIgnore if moized, originalFunction and options exist
+    return moize(fn.originalFunction, mergeOptions(fn.options, options));
+  }
 
-      return moize(fnOrOptions, {
-        // $FlowIgnore functionOrComposableOptions is object of options
-        ...functionOrComposableOptions,
-        ...otherOptions
-      });
+  if (typeof fn === 'object') {
+    return (curriedFn: Function | Options, curriedOptions: Options = {}) => {
+      return typeof curriedFn === 'function'
+        ? // $FlowIgnore fn is actually an object of options
+        moize(curriedFn, mergeOptions(fn, curriedOptions))
+        : // $FlowIgnore fn is actually an object of options
+        moize(mergeOptions(fn, curriedFn));
     };
   }
 
-  if (!isFunction(functionOrComposableOptions)) {
-    throw new TypeError(INVALID_FIRST_PARAMETER_ERROR);
-  }
-
-  const isComposed: boolean = functionOrComposableOptions.isMoized;
-  // $FlowIgnore if the function is already moized, it has an originalFunction property on it
-  const fn: Function = isComposed ? functionOrComposableOptions.originalFunction : functionOrComposableOptions;
-
-  const options: Options = getDefaultedOptions(
-    isComposed
-      ? {
-        ...functionOrComposableOptions.options,
-        ...passedOptions
-      }
-      : passedOptions
-  );
-
-  const cache: Cache = new Cache(options);
-
-  const addPropertiesToFunction: Function = createAddPropertiesToFunction(cache, fn, options);
-  const getCacheKey: Function = createGetCacheKey(cache, options);
-  const setNewCachedValue: Function = createSetNewCachedValue(cache, options);
-
-  return addPropertiesToFunction(function(...args: Array<any>): any {
-    const key: any = getCacheKey(args);
-
-    return cache.size && cache.has(key) ? cache.get(key) : setNewCachedValue(key, fn.apply(this, args));
+  const coalescedOptions: Options = Object.assign({}, DEFAULT_OPTIONS, options, {
+    profileName: options.profileName || getDefaultProfileName(fn)
   });
-};
+  const expirations: Array<Expiration> = [];
+
+  const {
+    equals: equalsIgnored,
+    isPromise,
+    isReact: isReactIgnored,
+    isSerialized: isSerialzedIgnored,
+    maxAge: maxAgeIgnored,
+    maxArgs: maxArgsIgnored,
+    maxSize,
+    onCacheAdd,
+    onCacheChange,
+    onCacheHit,
+    onExpire: onExpireIgnored,
+    profileName: profileNameIgnored,
+    shouldSerializeFunctions: shouldSerializeFunctionsIgnored,
+    serializer: serializerIgnored,
+    transformArgs: transformArgsIgnored,
+    updateExpire: updateExpireIgnored,
+    ...customOptions
+  } = coalescedOptions;
+
+  const isEqual: Function = getIsEqual(coalescedOptions);
+  const maxAgeOptions: Options = getMaxAgeOptions(expirations, coalescedOptions, isEqual);
+  const statsOptions: Options = getStatsOptions(coalescedOptions);
+  const transformKey: ?Function = getTransformKey(coalescedOptions);
+
+  const microMemoizeOptions: MicroMemoizeOptions = Object.assign({}, customOptions, {
+    isEqual,
+    isPromise,
+    maxSize,
+    onCacheAdd: createOnCacheOperation(combine(onCacheAdd, maxAgeOptions.onCacheAdd, statsOptions.onCacheAdd)),
+    onCacheChange: createOnCacheOperation(onCacheChange),
+    onCacheHit: createOnCacheOperation(combine(onCacheHit, maxAgeOptions.onCacheHit, statsOptions.onCacheHit)),
+    transformKey
+  });
+
+  return augmentMoizeInstance(memoize(fn, microMemoizeOptions), {
+    expirations,
+    options: coalescedOptions,
+    originalFunction: fn
+  });
+}
 
 /**
- * @function compose
+ * @function
+ * @name compose
+ * @memberof module:moize
+ * @alias moize.compose
  *
  * @description
  * method to compose moized methods and return a single moized function
@@ -119,15 +130,58 @@ const moize: Function = (functionOrComposableOptions: Function | Object, passedO
  * @param {...Array<(function)>} functions the functions to compose
  * @returns {function(...Array<*>): *} the composed function
  */
-moize.compose = compose;
+moize.compose = function() {
+  /* eslint-disable prefer-spread */
+  return compose.apply(null, arguments) || moize;
+  /* eslint-enable */
+};
 
 /**
- * @NOTE provide default export for TS to be able to import directly
+ * @function
+ * @name deep
+ * @memberof module:moize
+ * @alias moize.deep
+ *
+ * @description
+ * should deep equality check be used
+ *
+ * @returns {function} the moizer function
  */
-moize.default = moize;
+moize.deep = moize({isDeepEqual: true});
 
 /**
- * @function isMoized
+ * @function
+ * @name getStats
+ * @memberof module:moize
+ * @alias moize.getStats
+ *
+ * @description
+ * get the statistics of a given profile, or overall usage
+ *
+ * @returns {StatsProfile} statistics for a given profile or overall usage
+ */
+moize.getStats = getStats;
+
+/**
+ * @function
+ * @name isCollectingStats
+ * @memberof module:moize
+ * @alias moize.isCollectingStats
+ *
+ * @description
+ * are stats being collected
+ *
+ * @returns {boolean} are stats being collected
+ */
+moize.isCollectingStats = (): boolean => {
+  return statsCache.isCollectingStats;
+};
+
+/**
+ * @function
+ * @name isMoized
+ * @memberof module:moize
+ * @alias moize.isMoized
  *
  * @description
  * is the fn passed a moized function
@@ -136,100 +190,122 @@ moize.default = moize;
  * @returns {boolean} is fn a moized function
  */
 moize.isMoized = (fn: any): boolean => {
-  return isFunction(fn) && !!fn.isMoized;
+  return typeof fn === 'function' && !!fn.isMoized;
 };
 
 /**
- * @function maxAge
+ * @function
+ * @name maxAge
+ * @memberof module:moize
+ * @alias moize.maxAge
  *
  * @description
  * a moized method where the age of the cache is limited to the number of milliseconds passed
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @param {number} maxAge the TTL of the value in cache
+ * @returns {function} the moizer function
  */
-moize.maxAge = createCurriableOptionMethod(moize, 'maxAge');
+moize.maxAge = (maxAge: number): Function => {
+  return moize({maxAge});
+};
 
 /**
- * @function maxArgs
+ * @function
+ * @name maxArgs
+ * @memberof module:moize
+ * @alias moize.maxArgs
  *
  * @description
  * a moized method where the number of arguments used for determining cache is limited to the value passed
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @param {number} maxArgs the number of args to base the key on
+ * @returns {function} the moizer function
  */
-moize.maxArgs = createCurriableOptionMethod(moize, 'maxArgs');
+moize.maxArgs = (maxArgs: number): Function => {
+  return moize({maxArgs});
+};
 
 /**
- * @function maxSize
+ * @function
+ * @name maxSize
+ * @memberof module:moize
+ * @alias moize.maxSize
  *
  * @description
  * a moized method where the total size of the cache is limited to the value passed
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @param {number} maxSize the maximum size of the cache
+ * @returns {function} the moizer function
  */
-moize.maxSize = createCurriableOptionMethod(moize, 'maxSize');
+moize.maxSize = (maxSize: number): Function => {
+  return moize({maxSize});
+};
 
 /**
- * @function promise
+ * @function
+ * @name promise
+ * @memberof module:moize
+ * @alias moize.promise
  *
  * @description
  * a moized method specific to caching resolved promise / async values
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @returns {function} the moizer function
  */
 moize.promise = moize({
-  isPromise: true
+  isPromise: true,
+  updateExpire: true
 });
 
 /**
- * @function react
+ * @function
+ * @name react
+ * @memberof module:moize
+ * @alias moize.react
  *
  * @description
- * a moized method specific to caching React components
+ * a moized method specific to caching React element values
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @returns {function} the moizer function
  */
-moize.react = moize({
-  isReact: true
-});
+moize.react = moize({isReact: true});
 
 /**
- * @function reactSimple
+ * @function
+ * @name reactSimple
+ * @memberof module:moize
+ * @alias moize.reactSimple
  *
  * @description
- * a moized method specific to caching React components, only keeping the most recently-cached version
+ * a moized method specific to caching React element values, limiting to only the most recent result
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @returns {function} the moizer function
  */
-moize.reactSimple = compose(moize.react, moize.maxSize(1));
+moize.reactSimple = moize({isReact: true, maxSize: 1});
 
 /**
- * @function serialize
+ * @function
+ * @name serialize
+ * @memberof module:moize
+ * @alias moize.serialize
  *
  * @description
- * a moized method where the arguments passed are cached based on their serialized values
+ * a moized method that will serialize the arguments passed to use as the cache key
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @returns {function} the moizer function
  */
-moize.serialize = moize({
-  serialize: true
-});
+moize.serialize = moize({isSerialized: true});
 
 /**
- * @function simple
+ * @function
+ * @name simple
+ * @memberof module:moize
+ * @alias moize.simple
  *
  * @description
- * a moized method where only the most recent key => value combination is cached
+ * a moized method that will limit the cache values to only the most recent result
  *
- * @param {...Array<*>} functions the functions to compose
- * @returns {*} the moized function
+ * @returns {function} the moizer function
  */
 moize.simple = moize.maxSize(1);
 
